backend/requirements.txt
stripe
sqlalchemy
pydantic
langgraph
python-dotenv
langchain 
langchain_core 
langchain_community 
langchain_openai 
easyocr
numpy
pillow
uvicorn
fastapi
python-multipart
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib 